Refactor the data generation (#363)
This removes the circular dependency that made the app dependent on datafiles that were dependent on the app. Previously the app had to be built first, resource ids generated, the datafiles updated then the app rebuilt with the data files. This was so that the resource ids for all of the i18n strings could be properly embedded.

However it was super fragile. Add a resource and all the ids could change, causing mislabelled objects or, more often, crashes. Even different versions of the Android SDK seemed to be able to cause this. Ugly, brittle and dangerous to developers it had little to recommend it other than we knew no better at the time.

This is still quite ugly (the creation of a module to be included in the app containing the proto as well as the tools app reaching into that module to borrow a source file is not what I wanted), but should be less brittle. A missing resource will now result in an ugly "MISSING LABEL" label in the UI but should at least not crash.